pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autotool-system"
version = "0.1.0"
description = "Desktop automation toolkit."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pynput==1.7.6",
  "pyautogui==0.9.54",
  "pillow==10.0.0",
  "pyyaml==6.0.1",
  "schedule==1.2.1",
  "fastapi==0.115.6",
  "uvicorn==0.30.6",
]

[project.optional-dependencies]
dev = [
  "pytest==7.4.4",
  "pytest-cov==4.1.0",
  "requests==2.32.3",
  "httpx==0.27.2",
]
packaging = [
  "pyinstaller==6.3.0",
  "jaraco.functools==4.1.0",
  "jaraco.context==6.0.1",
  "jaraco.text==4.0.0",
]

[project.scripts]
autotool = "autotool_system.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["reference_project"]

[tool.coverage.run]
branch = true
source = ["src/autotool_system"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 60
